+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 In this changelog all modifications to the `citygml-convert-tools`
 
+## 2022-11-13 (v0.0.1)
+### Added
+* コマンドラインをサブコマンド形式に変更
+* PLYとGeoJSONの変換処理を統合
+* PLYとGeoJSONのテストを追加
+* earcut-pythonのサブモジュールを廃止し、プロジェクトに統合
+
+### Fixed
+* PLY変換のバグを修正
+
 ## 2022-11-05
 ### Added
 * citygml2geojsonを追加